refactor(ui): cleanup components
